+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#151](https://github.com/LaxarJS/laxar-uikit/issues/151): default.theme: added sass-options
+
 
 ## v2.0.0-beta.0